Update
1. 마을씬 다시 왔을 때 레벨업시 뻑나는 현상
2. HP, MP 자동 회복 구현

+++ Assets/KWS/Scripts/Base/Unit.cs
@@ -29,7 +29,7 @@ public class Unit : MonoBehaviour
     protected Weapon weapon;
     [SerializeField]
     protected UnitStats unit_stats;
-    protected float initial_mp;
+    protected int initial_mp;
     protected void Start()
     {
         unit_animation = GetComponent<Animator>();
